Move header up and move up to rmm:: namespace
