+++ tests/device_buffer_tests.cu
@@ -24,7 +24,7 @@
 #include <rmm/mr/device/managed_memory_resource.hpp>
 #include <rmm/mr/device/per_device_resource.hpp>
 #include <rmm/mr/device/pool_memory_resource.hpp>
-#include <rmm/mr/resource_ref.hpp>
+#include <rmm/resource_ref.hpp>
 
 #include <thrust/equal.h>
 #include <thrust/sequence.h>
@@ -59,7 +59,7 @@ struct DeviceBufferTest : public ::testing::Test {
 };
 
 using resources = ::testing::Types<rmm::mr::cuda_memory_resource, rmm::mr::managed_memory_resource>;
-using async_resource_ref = rmm::mr::device_async_resource_ref;
+using async_resource_ref = rmm::device_async_resource_ref;
 
 TYPED_TEST_CASE(DeviceBufferTest, resources);
