+++ tests/device_uvector_tests.cpp
@@ -18,7 +18,7 @@
 #include <rmm/cuda_stream_view.hpp>
 #include <rmm/device_uvector.hpp>
 #include <rmm/mr/device/per_device_resource.hpp>
-#include <rmm/mr/resource_ref.hpp>
+#include <rmm/resource_ref.hpp>
 
 #include <gtest/gtest.h>
 #include <gtest/internal/gtest-type-util.h>
@@ -32,7 +32,7 @@ struct TypedUVectorTest : ::testing::Test {
 };
 
 using TestTypes          = ::testing::Types<int8_t, int32_t, uint64_t, float, double>;
-using async_resource_ref = rmm::mr::device_async_resource_ref;
+using async_resource_ref = rmm::device_async_resource_ref;
 
 TYPED_TEST_CASE(TypedUVectorTest, TestTypes);
