+++ tests/mr/device/mr_ref_test.hpp
@@ -33,7 +33,7 @@
 #include <rmm/mr/device/owning_wrapper.hpp>
 #include <rmm/mr/device/per_device_resource.hpp>
 #include <rmm/mr/device/pool_memory_resource.hpp>
-#include <rmm/mr/resource_ref.hpp>
+#include <rmm/resource_ref.hpp>
 
 #include <gtest/gtest.h>
 
@@ -46,7 +46,7 @@
 #include <utility>
 
 using resource_ref       = cuda::mr::resource_ref<cuda::mr::device_accessible>;
-using async_resource_ref = rmm::mr::device_async_resource_ref;
+using async_resource_ref = rmm::device_async_resource_ref;
 
 namespace rmm::test {
